FindWestonClient.cmake: Update to detect that Weston is present.
We ensure Weston 6 or above is present as that is when xdg-stable
was merged.

The symbols which we used to rely on in libweston-desktop5 are no
longer present in libweston-desktop8 leading to undefined symbols
at runtime. Instead we include the generated wayland protocol files
directly and just use this cmake to ensure Weston is at least present
in the image.

Note: xdg-shell protocol is supported by many wayland compositors,
so in theory Thunder should work with mutt, gnome-shell etc, but for
now we leave it as a "Weston" thing, as that is all we tested it with.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>