compositorclient: add SurfaceByName() api
SurfaceByName() is used on the IDisplay class to get the ISurface.

It is used currently with Wayland Weston compositorclient when
using waylandsink as the video sink in webkit. wpebackend-rdk calls this
apis to obtain the surface handle created by Thunder, which is then passed to
waylandsink so that the video surface is correctly created as a sub-surface
of the browser.

Also RPI compositorclient is updated to have a stub implementation.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>