Wayland: Weston: remove xdg_toplevel_set_fullscreen()
If Weston is outputting at a higher resolution to what webkit plugin is
configured to in /etc/WPEFramework/plugins/WebKitBrowser.json then we get
a protocol error

xdg_surface buffer does not match the configured state.

With this patch applied, if 4k display is connected, and webkit is
confifgured to 1920x1080 then we will get browser in the appropriate
sized window.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>